backend-python/requirements.txt
Flask==2.3.3
boto3==1.40.38
python-dotenv==1.0.1
